requirements.txt
environs~=9.0
aiofiles==23.1.0
aiogram==3.0.0b7
python-dotenv==1.0.0
flask>=2.3.2
waitress>=2.1.2
betterlogging~=0.2.1
quart~=0.18.4
pymysql>=1.0.3